Add docker support & clarify onboarding setup
I'm lazy and don't want to have to install a bunch of stuff on my
machine, so I want to run the necessary services inside of Docker
containers.

This change adds support for this with docker composition. Currently,
only the MySQL container is added, but that's because I haven't been
able to figure out Tomcat yet. Once I get that working, I'll make the
necessary changes to `docker-compose.yml` and make a new commit.

For now, a simple `docker compose up` will get the dev environment
running (though you'll need to set up the database by feeding in the
`init.sql` script, too).